server/requirements.txt
fastapi[standard]
uvicorn
pydantic-settings
sqlalchemy
psycopg2
unstructured[pdf]
unstructured-inference
langchain
langchain-openai
langchain-community
pinecone
openai
python-dotenv
requests
aiofiles
pycryptodome
pdf2image
pillow
pytesseract
